(FACT-613) Sync flat processor facts with structured fact
This commit refactors the Processor, ProcessorCount,
PhysicalProcessorCount and dynamic ProcessorX facts to
resolve by using values from the new 'processors' structured
fact. This ensures that these facts will stay in sync with
the new fact, while also maintaining backwards compatibility.
